fix(cloister): preserve specialist seam privacy (PAN-2232)
